+++ ChangeLog.md
@@ -6,6 +6,7 @@ See also http://pvp.haskell.org/faq
 - New `qnameToText` and `qnameFromText` conversion functions
 - Explicitly declare SafeHaskell levels; ensure all public modules are "Safe"
 - Fix pure exception bypassing proper failure reporting
+- Fix attribute uniqueness violations not being detected
 
 ## 0.3.0.0